// src/app/shared/types/status.type.ts
export declare type StatusTypes =
    | 'All Active'
    | 'At Origin (Job Added)'
    | 'Departed'
    | 'Freight Check in at Origin'
    | 'Awaiting Arrival at Destination'
    | 'Customs Clearance'
    | 'At Transit Port'
    | 'In Transit to Arrival'
    | 'Destination Stop-off'
    | 'At final Destination'
    | 'Invoices not settled'
    | 'Completed';